restructured into packages + reset GUI at reconnection attempt
