+++ src/main/java/at/fhhagenberg/sqe/ElevatorControl.java
@@ -1,25 +1,18 @@
 package at.fhhagenberg.sqe;
 
 import java.net.MalformedURLException;
-import java.rmi.Naming;
 import java.rmi.NotBoundException;
 import java.rmi.RemoteException;
-import java.util.Locale;
-
-import javax.swing.event.ChangeEvent;
-import javax.swing.event.ChangeListener;
-
-import at.fhhagenberg.sqe.controlcenter.BuildingAdapter;
 import at.fhhagenberg.sqe.controlcenter.ControlCenterException;
-import at.fhhagenberg.sqe.controlcenter.IBuilding;
-import at.fhhagenberg.sqe.controlcenter.mocks.BuildingMock;
+import at.fhhagenberg.sqe.controller.ElevatorControlController;
+import at.fhhagenberg.sqe.exceptionhandler.ElevatorExceptionHandler;
+import at.fhhagenberg.sqe.exceptionhandler.RemoteElevatorExceptionHandler;
 import at.fhhagenberg.sqe.model.BuildingModel;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
-import sqelevator.IElevator;
 
 /**
  * JavaFX App
