+++ src/main/java/at/fhhagenberg/sqe/controlcenter/mocks/ElevatorControlMock.java
@@ -2,10 +2,8 @@ package at.fhhagenberg.sqe.controlcenter.mocks;
 
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
 
-import at.fhhagenberg.sqe.controlcenter.ControlCenterException;
 import at.fhhagenberg.sqe.controlcenter.IElevatorControl;
 
 public class ElevatorControlMock implements IElevatorControl {
