+++ src/main/resources/at/fhhagenberg/sqe/controller/ElevatorFloor.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.layout.AnchorPane?>
 
-<HBox id="elevatorFloorHBox" fx:id="elevatorFloorHBox" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="40.0" minWidth="115.0" prefHeight="40.0" prefWidth="115.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="at.fhhagenberg.sqe.ElevatorFloorController">
+<HBox id="elevatorFloorHBox" fx:id="elevatorFloorHBox" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="40.0" minWidth="115.0" prefHeight="40.0" prefWidth="115.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="at.fhhagenberg.sqe.controller.ElevatorFloorController">
    <children>
       <ImageView id="elevatorImageView" fx:id="elevatorImageView" fitHeight="40.0" fitWidth="32.0" pickOnBounds="true" preserveRatio="true">
          <image>
