+++ src/main/resources/at/fhhagenberg/sqe/controller/elevator.png
