+++ src/main/resources/at/fhhagenberg/sqe/controller/output-onlinepngtools-2.png
