+++ src/main/resources/at/fhhagenberg/sqe/controller/output-onlinepngtools.png
