+++ src/test/java/at/fhhagenberg/sqe/ElevatorControlEndToEndTest.java
@@ -19,6 +19,8 @@ import org.testfx.matcher.control.TextInputControlMatchers;
 import at.fhhagenberg.sqe.controlcenter.BuildingAdapter;
 import at.fhhagenberg.sqe.controlcenter.ControlCenterException;
 import at.fhhagenberg.sqe.controlcenter.mocks.BuildingMock;
+import at.fhhagenberg.sqe.controller.ElevatorControlController;
+import at.fhhagenberg.sqe.exceptionhandler.RemoteElevatorExceptionHandler;
 import at.fhhagenberg.sqe.model.BuildingModel;
 import javafx.stage.Stage;
 import sqelevator.IElevator;
