+++ src/test/java/at/fhhagenberg/sqe/ElevatorControlGuiTest.java
@@ -13,6 +13,7 @@ import org.testfx.matcher.control.ListViewMatchers;
 
 import at.fhhagenberg.sqe.controlcenter.ControlCenterException;
 import at.fhhagenberg.sqe.controlcenter.mocks.BuildingMock;
+import at.fhhagenberg.sqe.controller.ElevatorControlController;
 import at.fhhagenberg.sqe.model.BuildingModel;
 import javafx.stage.Stage;
