+++ src/test/java/at/fhhagenberg/sqe/controller/ElevatorControllerTest.java
@@ -1,4 +1,4 @@
-package at.fhhagenberg.sqe;
+package at.fhhagenberg.sqe.controller;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -15,6 +15,7 @@ import at.fhhagenberg.sqe.controlcenter.IElevatorControl.Direction;
 import at.fhhagenberg.sqe.controlcenter.IElevatorControl.DoorStatus;
 import at.fhhagenberg.sqe.controlcenter.mocks.BuildingMock;
 import at.fhhagenberg.sqe.controlcenter.mocks.ElevatorControlMock;
+import at.fhhagenberg.sqe.controller.ElevatorController;
 import at.fhhagenberg.sqe.model.BuildingModel;
 import at.fhhagenberg.sqe.model.ElevatorModel;
